Submodularized RefineSyntax modules + little code cleaning
