+++ frank.cabal
@@ -19,6 +19,9 @@ executable frank
                , FreshNames
                , Parser
                , RefineSyntax
+               , RefineSyntaxCommon
+               , RefineSyntaxConcretiseEps
+               , RefineSyntaxSubstitItfAliases
                , Shonky.Semantics
                , Shonky.Syntax
                , Syntax
@@ -42,7 +45,7 @@ executable frank
                , unordered-containers
                , cmdargs
   main-is:          Frank.hs
-  ghc-options:      -threaded -rtsopts -with-rtsopts=-N 
+  ghc-options:      -threaded -rtsopts -with-rtsopts=-N
   default-language: Haskell2010
 
 source-repository head
